changes in attack phase
