+++ src/com/riskgame/model/Player.java
@@ -105,26 +105,6 @@ public class Player {
 		this.myCountries.add(country);
 	}
 
-	public void attackPhase(Country attacker, Country defender) {
-		boolean isAttackPossible = false;
-		if(attacker != null && defender != null) {
-			if(attacker.getAdjacentCountries().contains(defender.getName())) {
-				if(attacker.getNoOfArmies() > 1 && defender.getNoOfArmies() > 0) {
-					isAttackPossible = true;
-				}
-				else {
-					System.out.println("Insufficient armies in the attacker country/defender country");
-				}
-			}
-			else {
-				System.out.println("Attacker and Defender Countries are not adjacent!");
-			}	
-		}
-		
-		if(isAttackPossible) {
-			//attacker and defender need to select the number of dice to roll
-		}
-	}
 	
 	@Override
 	public String toString() {
